refactor(trace): build the trace fixtures once per crate
